+++ Controllers/AccountController.cs
@@ -1,15 +1,20 @@
 
 using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Authentication;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.Extensions.Logging;
 using VidsNet.DataModels;
+using VidsNet.Enums;
 using VidsNet.Interfaces;
 using VidsNet.Models;
+using VidsNet.Scanners;
 using VidsNet.ViewModels;
 
 namespace VidsNet.Controllers
@@ -19,11 +24,13 @@ namespace VidsNet.Controllers
     {
         private IUserRepository _userRepository;
         private readonly ILogger _logger;
-        public AccountController(IUserRepository userRepository, ILoggerFactory loggerFactory, UserData userData)
+        private BaseScanner _baseScanner;
+        public AccountController(IUserRepository userRepository, ILoggerFactory loggerFactory, UserData userData, BaseScanner baseScanner) 
          : base(userData)
         {
             _userRepository = userRepository;
             _logger = loggerFactory.CreateLogger("AccountController");
+            _baseScanner = baseScanner;
         }
 
         /*[HttpPost]
@@ -93,9 +100,96 @@ namespace VidsNet.Controllers
             }
             return View(new LoginViewModel(_user){ ErrorMessage = "Unknown error."});
         }
+
+        private void Helper(Item item, TagBuilder select, int level, List<UserSetting> userPaths) {
+            var padding = level * 25;
+            foreach (var child in item.Children)
+            {
+                var guid = Guid.NewGuid().ToString();
+                var div = new TagBuilder("div");
+                div.AddCssClass("list-group-item");
+                div.MergeAttribute("style", string.Format("padding-left: {0}px", padding));
+
+                var checkbox = new TagBuilder("input");
+                checkbox.MergeAttribute("type", "checkbox");
+                checkbox.MergeAttribute("name", child.Path);
+                checkbox.MergeAttribute("value", child.Path);
+
+                if(userPaths.Any(x => x.Value == string.Format("{0}{1}", child.Path, Path.DirectorySeparatorChar))) {
+                    checkbox.MergeAttribute("checked", "checked");
+                }
+
+
+                div.InnerHtml.AppendHtml(checkbox);
+
+                if(child.Children.Count > 0) {
+                    var button = new TagBuilder("button");
+                    button.AddCssClass("btn");
+                    button.AddCssClass("btn-link");
+                    button.MergeAttribute("data-toggle", "collapse");
+                    
+                    button.MergeAttribute("data-target", string.Format("#{0}", guid));
+                    button.MergeAttribute("type", "button");
+
+                    var nameSpan = new TagBuilder("span");
+                    nameSpan.InnerHtml.Append(child.Path);
+                    nameSpan.MergeAttribute("id", string.Format("{0}_name", guid));
+                    nameSpan.MergeAttribute("style", "margin-left: 15px;");
+
+                    var buttonSpan = new TagBuilder("span");
+                    buttonSpan.AddCssClass("glyphicon");
+                    buttonSpan.AddCssClass("glyphicon-folder-open");
+                    buttonSpan.MergeAttribute("aria-hidden", "true");
+                    buttonSpan.InnerHtml.AppendHtml(nameSpan);
+
+                    button.InnerHtml.AppendHtml(buttonSpan);
+                    div.InnerHtml.AppendHtml(button);
+                }
+                else {
+                    var nameSpan = new TagBuilder("span");
+                    nameSpan.InnerHtml.Append(child.Path);
+                    nameSpan.MergeAttribute("id", string.Format("{0}_name", guid));
+                    nameSpan.MergeAttribute("style", "margin-left: 15px;");
+
+                    var buttonSpan = new TagBuilder("span");
+                    buttonSpan.AddCssClass("glyphicon");
+                    buttonSpan.AddCssClass("glyphicon-folder-close");
+                    buttonSpan.MergeAttribute("aria-hidden", "true");
+                    buttonSpan.MergeAttribute("style", "padding: 6px 12px");
+                    buttonSpan.InnerHtml.AppendHtml(nameSpan);
+                    div.InnerHtml.AppendHtml(buttonSpan);
+                }
+
+                select.InnerHtml.AppendHtml(div);
+
+                var collapsedDiv = new TagBuilder("div");
+                collapsedDiv.MergeAttribute("id", guid);
+                collapsedDiv.AddCssClass("collapse");
+                
+
+                if(child.Type == ItemType.Folder) {
+                    Helper(child, collapsedDiv, level + 1, userPaths);
+                }
+                select.InnerHtml.AppendHtml(collapsedDiv);
+            }
+        }
+
         [HttpGet]
         public IActionResult Settings() {
-            var settings = new SettingsViewModel() { IsAdmin = _user.IsAdmin, UserSettings = _user.UserSettings, AdminSettings = _user.AdminSettings };
+            var home = Environment.GetEnvironmentVariable("HOME");
+            var folders = _baseScanner.ScanItems(new List<string>() { home }, new List<IScannerCondition>());
+            var userPaths = _user.UserSettings.Where(x => x.Name == "path").ToList();
+            var form = new TagBuilder("form");
+            form.MergeAttribute("id", "userPaths");
+
+            foreach(var folder in folders) {
+                Helper(folder, form, 1, userPaths);
+            }
+            var settings = new SettingsViewModel(_user) { 
+                IsAdmin = _user.IsAdmin, 
+                UserSettings = _user.UserSettings, 
+                AdminSettings = _user.AdminSettings,
+                DirectoryListing = form };
             return View(settings);
         }