+++ DataModels/Item.cs
@@ -1,7 +1,5 @@
-
-
 using System;
-using System.Collections;
+using System.Collections.Generic;
 using VidsNet.Enums;
 
 namespace VidsNet.DataModels
@@ -10,17 +8,25 @@ namespace VidsNet.DataModels
 
         public string Path { get; set; } 
         public int Id { get; set; }
+        public int UserPathId {get;set;}
         public ItemType Type {get;set;}
+        public bool WriteVirtualItem {get;set;}
+        public List<Item> Children {get;set;}
+
+        public Item () {
+            Children = new List<Item>();
+            WriteVirtualItem = true;
+        }
 
 
         public override int GetHashCode()
         {
-            return string.Format("{0}.{1}.{2}", Id, Path, Type).GetHashCode();
+            return string.Format("{0}.{1}", Path, Type).GetHashCode();
         }
 
         bool IEquatable<Item>.Equals(Item other)
         {
-            if(other.Id == Id) {
+            if(other.Path.Equals(Path)) {
                 return true;
             }
             return false;