+++ Interfaces/IScannerCondition.cs
@@ -1,0 +1,8 @@
+using VidsNet.DataModels;
+
+namespace VidsNet.Interfaces
+{
+    public interface IScannerCondition {
+        CheckTypeResult CheckType(string filePath);
+    }
+}