+++ ViewModels/BaseViewModel.cs
@@ -1,6 +1,7 @@
 using VidsNet.Classes;
 using System.Linq;
 using VidsNet.DataModels;
+using System;
 
 namespace VidsNet.ViewModels
 {